Dataset : MNIST, Purpose : Image Classification
The MNIST (Modified National Institute of Standards and Technology) is a large database of handwritten digits that is commonly used for training various image processing systems. Here, the Dataset used is EMNIST's 'Digits'. (EMNIST stands for Extended MNIST).

A CNN is a Neural Network: An Algorithm used to recognize patterns in data. Neural Networks in general are composed of a collection of neurons that are organized in layers, each with their own learnable weights and biases.

Here, The Network Architecture used is TinyVGG. Resource : https://poloclub.github.io/cnn-explainer/

Number of Model Parameters : 7740 Parameters
Version 1

More on the Dataset : https://en.wikipedia.org/wiki/MNIST_database